requirements.txt
langchain
langchain_community
python-dotenv
crewai_tools
sqlite3
ollama
